==> 2_Epidemiological_data.Rmd <==
---
title: "Epidemiological data"
author: 
- name: "Jason_Rosado"
affiliation: Institut de Recherche pour le Développement
email: jason.rosado@ird.fr 
date: "`r Sys.Date()`"
output: 
  html_document:
    css: mystyle.css
    number_sections: yes
    toc: true
    toc_float: true
    code_folding: hide 
    code_download: false
    fig_width: 8
    fig_height: 6
    theme: cosmo
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      tidy = TRUE,
                      fig.width = 8, 
                      fig.height = 6)

library(plyr)
library(table1)
library(tidyverse)
library(table1)
library(miplicorn)
library(ggplot2)
library(cowplot)
library(ggpubr)
library(corrplot)
library(ggcorrplot)
library(forcats)
library(tableone)
library(here)
library(rnaturalearth)
library(ggspatial)
  
```

#Loading data from previous runs
```{r, include=FALSE}
IC50_genotype <- readRDS(here("intermediate_data", "IC50_genotype.rds"))
```

# Epi table
```{r,echo=F}

#Function to transform numeric to categorical variable

rndr <- function(x, ...) {
    if (is.factor(x) || is.character(x)) {
        c(render.default(x, ...), c(`Overall N`=sum(!is.na(x))))
    } else {
        render.default(x, ...)
    }
}


table1(~  age +  gender + parasitemia + median_coi + DHA+ 
         Lumefantrine + Mefloquine + MDAQ + chloroquine +
        piperaquine + country 
     , data = IC50_genotype,render.continuous = c("Median [Q1, Q3]") )




```
```{r, echo=FALSE}
COI_cat<- IC50_genotype%>%
  dplyr::select(median_coi) %>%
  mutate(coi_cat = ifelse(median_coi>1, "1+", "1")) 

table1(~  coi_cat 
     , data = COI_cat,render.continuous = c("Median [Q1, Q3]") )
```

```{r, echo=FALSE, message=FALSE}
ranking_country <- IC50_genotype %>%
  group_by(country) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  mutate(freq = n/sum(n)) %>%
  arrange(desc(freq))

write.csv(ranking_country, "tables/ranking_country.csv", row.names = F)

```


Countries with frequency >0.03%
```{r, echo=FALSE, message=FALSE}

top_countries <- IC50_genotype %>%
  rename(admin = country) %>%
  group_by(admin) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  mutate(Frequency = n/sum(n)) %>%
   arrange(desc(Frequency)) %>%
    filter(Frequency>=0.003) 


africa <- ne_countries(scale="medium", type = "sovereignty", continent = "Africa", returnclass = "sf", )

top_countries_map<- africa %>%
  dplyr::right_join(top_countries, by = "admin")



# Africa vs Top countries Map


colors <- c("#FFFFFF","#FFFF33", "#FF8000", "#FF0000", "#990000")
            


country_map<- ggplot() + 
  geom_sf(data=africa, fill="gray90")+
  geom_sf(data = top_countries_map, aes(fill= n), 
          colour = "black") +
  scale_fill_stepsn(colors = colors, 
                    breaks = c(3,20,50,100),
                    labels = c("3-20","21-50","51-100", "100+"),
                    values = scales::rescale(c(3,20,50,100)),
                    name = "Number of isolates") +
  theme_nothing() +  
  theme(legend.position = c(.2,.4)) 
 
print(country_map)
ggsave("figures/country_map.pdf", units="cm", width=19, height=15, dpi=300)

```

### Years in CNR database
```{r, echo=FALSE, message=FALSE}
ranking_year <- IC50_genotype %>%
  group_by(year) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  mutate(freq = n/sum(n)) %>%
  arrange(desc(freq))

top_countries_year <- IC50_genotype %>%
  left_join(ranking_country, by = "country") %>%
  rename(admin = country) %>%
  mutate(admin = case_when(admin =="Ivory Coast" ~ "Côte_d_Ivoire",
    freq<"0.027" ~ "n < 20",
                               TRUE ~ as.character(admin))) %>%
  dplyr::select(year, admin) %>%
  group_by(year, admin) %>%
  summarise(n = n()) %>%
  ungroup() %>%
  mutate(Frequency = n/sum(n)) %>%
  arrange(desc(Frequency)) %>%
  mutate(year = as.factor(year))
  

top_countries_year$admin <-reorder(top_countries_year$admin, ifelse(!top_countries_year$admin %in% "Côte_d_Ivoire",0, top_countries_year$n), Fun = sum) 

top_countries_year_plot<-  ggplot(data =top_countries_year, aes(year, n, 
                                       fill = factor(admin, 
                                  levels = c("n < 20", "Gabon",
                                           "Benin","Chad","Central African Republic",                                                 "Senegal","Republic of the Congo",
                                            "Guinea","Mali","Cameroon","Côte_d_Ivoire")))) +
  geom_bar(position = "stack", stat = "identity", color = "black") +
    scale_fill_brewer(palette = "Set3") +
  ggtitle("") + 
  ylab("Number of isolates") +
  xlab("year") + 
  scale_x_discrete( breaks=c("2016","2017", "2017","2018","2019","2020", "2021", "2022","2023"),
                    labels=c("2016","2017", "2017","2018","2019","2020", "2021", "2022","2023")) + ylim(0, 200) +
  theme_classic(base_size = 30) + 
  guides(fill = guide_legend(title = "")) +
  theme(axis.text.x = element_text(angle=45, vjust=0.5, size = 25))

print(top_countries_year_plot)
ggsave("figures/top_countries_year_plot.pdf", units="cm", width=28, height=22, dpi=300)

```


```{r echo=F, eval=F}
save.image("./intermediate_data/epidemiological_data.RData")
```
